Simplified the entire setup for using classes for mnanaging the user records.
