--- src/main/java/runner/H2DBInterface.java
@@ -1,18 +1,0 @@
-package runner;
-
-/**
- * Created by Don Robles on 2/4/2017 with IntelliJ IDEA.
- * <p>
- * This interface contains the base SQL statements that the H2 DB will use.
- */
-public interface H2DBInterface {
-
-    String DROP_TABLES = "DROP TABLE IF EXISTS EMAIL_ADDRESSES, USER_INFO, PHONE_NUMBERS;";
-    String USER_INFO = "CREATE TABLE USER_INFO (USER_ID INTEGER AUTO_INCREMENT PRIMARY KEY, FIRST_NAME VARCHAR(45), LAST_NAME VARCHAR(45));";
-    String PHONE_NUMBERS =
-            "CREATE TABLE PHONE_NUMBERS (PHONE_NUMBER_ID INTEGER AUTO_INCREMENT PRIMARY KEY, PHONE_NUMBER VARCHAR(12)," +
-                    "CONSTRAINT PHONE_NUMBERS_USER_INFO_USER_ID_FK FOREIGN KEY (PHONE_NUMBER_ID) REFERENCES USER_INFO (USER_ID))";
-    String EMAIL_ADDESSES =
-            "CREATE TABLE EMAIL_ADDRESSES (EMAIL_ADDRESS_ID INTEGER AUTO_INCREMENT PRIMARY KEY, EMAIL_ADDRESS VARCHAR(45)," +
-                    "CONSTRAINT EMAIL_ADDRESSES_USER_INFO_USER_ID_FK FOREIGN KEY (EMAIL_ADDRESS_ID) REFERENCES USER_INFO (USER_ID));";
-}
